+++ packages/expo-modules-core/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### 🎉 New features
 
 - Add `preventModuleOverriding` to `ModuleRegistry.register` method. ([#24860](https://github.com/expo/expo/pull/24860) by [@wschurman](https://github.com/wschurman))
+- [iOS] `CGFloat` can now be used as an argument type. ([#25140](https://github.com/expo/expo/pull/25140) by [@tsapeta](https://github.com/tsapeta))
 
 ### 🐛 Bug fixes